Move code from `app.js` to `index.js`
